fix: isolate design runtime verification
