+++ sdd/spec/changes.md
@@ -4,7 +4,7 @@ Semantic changes to the specification. Git history captures diffs; this file cap
 
 ## 2026-08-13
 
-- **Advanced sessions gain one master design router and three portable specialists** ([REQ-AGENT-134](agents.md#req-agent-134-advanced-design-skill-suite) and [REQ-AGENT-135](agents.md#req-agent-135-ui-ux-pro-max-runtime-contract) added; Partial pending review and CI). The canonical Claude preseed now carries a compact `design` entry point that composes UI UX Pro Max for searchable product-interface guidance, Canvas Design for static PNG/PDF artwork, an independently written Codeflare Frontend Design skill for distinctive frontend direction, and optional Impeccable critique where that full tool bundle is installed. The standard generator transforms the suite for Codex, Antigravity, OpenCode, and Pi; Copilot remains excluded because it has no skill-directory contract. UI UX Pro Max retains its MIT license, Canvas Design retains Apache-2.0 with provenance and modification notice, and no all-rights-reserved Anthropic Claude Code source is copied or transformed.
+- **Advanced sessions gain one master design router and three portable specialists** ([REQ-AGENT-134](agents.md#req-agent-134-advanced-design-skill-suite) and [REQ-AGENT-135](agents.md#req-agent-135-ui-ux-pro-max-query-and-generation), [REQ-AGENT-136](agents.md#req-agent-136-ui-ux-pro-max-persistence), and [REQ-AGENT-137](agents.md#req-agent-137-ui-ux-pro-max-data-validation) added; Partial pending review and CI). The canonical Claude preseed now carries a compact `design` entry point that composes UI UX Pro Max for searchable product-interface guidance, Canvas Design for static PNG/PDF artwork, an independently written Codeflare Frontend Design skill for distinctive frontend direction, and optional Impeccable critique where that full tool bundle is installed. The standard generator transforms the suite for Codex, Antigravity, OpenCode, and Pi; Copilot remains excluded because it has no skill-directory contract. UI UX Pro Max retains its MIT license, Canvas Design retains Apache-2.0 with provenance and modification notice, and no all-rights-reserved Anthropic Claude Code source is copied or transformed.
 
 - **Native Codeflare Chat keeps one lazy IDE-owned Pi conversation across panel and editor turns** ([REQ-IDE-005](browser-ide.md#req-ide-005-selected-native-ide-agent), [REQ-IDE-006](browser-ide.md#req-ide-006-ide-conversation-context-and-credential-isolation), [REQ-IDE-008](browser-ide.md#req-ide-008-ide-agent-process-lifecycle), [REQ-IDE-019](browser-ide.md#req-ide-019-codeflare-eligibility-in-editor-inline-chat), [REQ-IDE-020](browser-ide.md#req-ide-020-unrestricted-pi-editor-request-execution), and [REQ-IDE-022](browser-ide.md#req-ide-022-native-pi-blocking-ui-protocol) amended; affected runtime requirements Partial pending review and CI). The first request lazily starts the fixed unrestricted local Pi RPC process, and normally completed turns retain it. Panel and editor requests execute FIFO and intentionally share its hidden in-memory transcript, while terminal Pi remains a separate process and conversation. Cold creation or replacement bootstraps from the requesting surface's bounded visible history; warm turns send only current request/editor context to avoid replaying history Pi already holds. Queued cancellation skips only that request. Active cancellation, startup cancellation after backend creation, protocol or process failure, unexpected exit, and deactivation boundedly retire the generation before replacement. No code-server, Code OSS, Docker pin, Bump Shadow Pins workflow, tool, root, interceptor, approval, or fail-closed protocol boundary changes.
